refactor: Reorganize authentication and security components into a new `auth` package structure.
