+++ backend/src/main/java/com/swipelab/auth/application/EmailService.java
@@ -1,4 +1,4 @@
-package com.swipelab.service.auth;
+package com.swipelab.auth.application;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
