+++ backend/src/main/java/com/swipelab/auth/infrastructure/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package com.swipelab.security;
+package com.swipelab.auth.infrastructure;
 
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
