+++ backend/src/main/java/com/swipelab/auth/infrastructure/OAuth2Config.java
@@ -1,0 +1,5 @@
+package com.swipelab.auth.infrastructure;
+
+public class OAuth2Config {
+    
+}
