+++ backend/src/main/java/com/swipelab/auth/infrastructure/OAuth2SuccessHandler.java
@@ -1,4 +1,4 @@
-package com.swipelab.security;
+package com.swipelab.auth.infrastructure;
 
 public class OAuth2SuccessHandler {
