--- backend/src/main/java/com/swipelab/config/OAuth2Config.java
@@ -1,5 +1,0 @@
-package com.swipelab.config;
-
-public class OAuth2Config {
-    
-}
